ai-agent/requirements.txt
google-generativeai
elasticsearch